+++ src/main/java/umc/unimade/domain/products/controller/ProductsController.java
@@ -1,4 +1,4 @@
-package umc.unimade.domain.products;
+package umc.unimade.domain.products.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
