+++ src/main/java/umc/unimade/domain/products/service/ProductsCommandService.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.unimade.domain.accounts.entity.Buyer;
-import umc.unimade.domain.accounts.entity.Seller;
 import umc.unimade.domain.accounts.repository.BuyerRepository;
 import umc.unimade.domain.favorite.entity.FavoriteProduct;
 import umc.unimade.domain.favorite.repository.FavoriteProductRepository;
@@ -14,7 +13,7 @@ import umc.unimade.domain.products.entity.Category;
 import umc.unimade.domain.products.entity.ProductRegister;
 import umc.unimade.domain.products.repository.CategoryRepository;
 import umc.unimade.domain.products.repository.OptionsRepository;
-import umc.unimade.domain.products.ProductRepository;
+import umc.unimade.domain.products.repository.ProductRepository;
 import umc.unimade.domain.products.entity.Products;
 import umc.unimade.domain.products.repository.ProductRegisterRepository;
 import umc.unimade.global.common.ApiResponse;
